중간 마지막 업데이트

+++ src/main/java/com/example/demo/BookDeleteService.java
@@ -4,10 +4,10 @@ import java.util.Scanner;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+//등록되 있는 책을 삭제하는 기능
 public class BookDeleteService {
 
 	private BookDao bookDao;
-//	BookRegisterRequest req  = new BookRegisterRequest();
 	Scanner sc = new Scanner(System.in);
 
 	private String bookISBN;
@@ -25,16 +25,28 @@ public class BookDeleteService {
 		
 		System.out.print("도서 ISBN > ");
 		bookISBN = sc.nextLine();
-//		req.setBookISBN(bookISBN);
+	
+		Book book = null;
+		book = bookDao.selectByBookISBN(bookISBN);
 		
 		System.out.print("정말로 삭제하시겠습니까? (Y 입력시 삭제) > ");
 		confirm = sc.nextLine();
-		if (confirm.equals("Y")) {
-			bookDao.remove(bookISBN);
-			System.out.print("도서를 삭제하였습니다.");
+		
+		//해당 ISBN 을 가진 도서가 존재하는 경우
+		if (book!=null){
+			// 정말로 삭제하겠다고 한 경우
+			if (confirm.equals("Y")) {
+				bookDao.remove(bookISBN);
+				System.out.print("도서를 삭제하였습니다.");
+			}
+			//취소했을 경우
+			else if(!confirm.equals("Y")) {
+				System.out.print("도서 삭제를 취소하셨습니다.");
+			}
 		}
+		//해당 ISBN 을 가진 도서가 존재하지 않는 경우
 		else {
-			System.out.print("도서 삭제를 취소하셨습니다.");
+			System.out.println("해당 ISBN 을 가진 도서는 존재하지 않습니다.");
 		}
 	}
